2026-01-10-google-embedding-mvp/requirements.txt
google-generativeai
python-dotenv
streamlit
pypdf
Pillow
